Change code as suggested in PR
